Replace elastic search entities and classes to data layer
