+++ OnlinerTask/OnlinerTask.BLL/Services/Logger/ILogger.cs
@@ -1,0 +1,7 @@
+﻿namespace OnlinerTask.BLL.Services.Logger
+{
+    public interface ILogger
+    {
+        void LogObject(object logEntity);
+    }
+}
