+++ OnlinerTask/OnlinerTask.BLL/Services/Logger/Implementations/Logger.cs
@@ -1,0 +1,10 @@
+﻿namespace OnlinerTask.BLL.Services.Logger.Implementations
+{
+    public class Logger : ILogger
+    {
+        public void LogObject(object logEntity)
+        {
+
+        }
+    }
+}
