+++ OnlinerTask/OnlinerTask.Data/ElasticSearch/ProductLogger/Implementations/ProductAddLogger.cs
@@ -1,26 +1,26 @@
 ﻿using System.Threading.Tasks;
 using Nest;
-using OnlinerTask.BLL.Services.ElasticSearch.ClientsFabric;
-using OnlinerTask.BLL.Services.ElasticSearch.ConnectionFabric;
-using OnlinerTask.Data.DataBaseModels;
+using OnlinerTask.Data.ElasticSearch.ClientsFabric;
+using OnlinerTask.Data.ElasticSearch.ConnectionFabric;
 using OnlinerTask.Data.Resources;
+using OnlinerTask.Data.SearchModels;
 
-namespace OnlinerTask.BLL.Services.ElasticSearch.ProductLogger.Implementations
+namespace OnlinerTask.Data.ElasticSearch.ProductLogger.Implementations
 {
-    public class ProductRemoveLogger : IProductLogger<Product>
+    public class ProductAddLogger : IProductLogger<ProductModel>
     {
         private readonly ElasticClient elasticClient;
 
-        public ProductRemoveLogger(IClientsFactory clientsFabric, IConnectionFactory connectionFabric)
+        public ProductAddLogger(IClientsFactory clientsFabric, IConnectionFactory connectionFabric)
         {
             var settings = connectionFabric.CreateConnection(Configurations.ElasticSearchUrl, Configurations.ProductLogIndex);
             elasticClient = clientsFabric.CreateClient(settings, Configurations.ProductLogIndex);
         }
 
-        public async Task LogObject(Product productModel)
+        public async Task LogObject(ProductModel productModel)
         {
             var defaultIndex = elasticClient.ConnectionSettings.DefaultIndex;
-            var indexType = Configurations.ProductRemoveIndexType;
+            var indexType = Configurations.ProductAddIndexType;
 
             await elasticClient.IndexAsync(productModel,
                 i => i.Index(defaultIndex).Type(indexType).Refresh());
