+++ OnlinerTask/OnlinerTask.Data/ElasticSearch/ProductLogger/Implementations/ProductRemoveLogger.cs
@@ -1,26 +1,26 @@
 ﻿using System.Threading.Tasks;
 using Nest;
-using OnlinerTask.BLL.Services.ElasticSearch.ClientsFabric;
-using OnlinerTask.BLL.Services.ElasticSearch.ConnectionFabric;
+using OnlinerTask.Data.DataBaseModels;
+using OnlinerTask.Data.ElasticSearch.ClientsFabric;
+using OnlinerTask.Data.ElasticSearch.ConnectionFabric;
 using OnlinerTask.Data.Resources;
-using OnlinerTask.Data.SearchModels;
 
-namespace OnlinerTask.BLL.Services.ElasticSearch.ProductLogger.Implementations
+namespace OnlinerTask.Data.ElasticSearch.ProductLogger.Implementations
 {
-    public class ProductAddLogger : IProductLogger<ProductModel>
+    public class ProductRemoveLogger : IProductLogger<Product>
     {
         private readonly ElasticClient elasticClient;
 
-        public ProductAddLogger(IClientsFactory clientsFabric, IConnectionFactory connectionFabric)
+        public ProductRemoveLogger(IClientsFactory clientsFabric, IConnectionFactory connectionFabric)
         {
             var settings = connectionFabric.CreateConnection(Configurations.ElasticSearchUrl, Configurations.ProductLogIndex);
             elasticClient = clientsFabric.CreateClient(settings, Configurations.ProductLogIndex);
         }
 
-        public async Task LogObject(ProductModel productModel)
+        public async Task LogObject(Product productModel)
         {
             var defaultIndex = elasticClient.ConnectionSettings.DefaultIndex;
-            var indexType = Configurations.ProductAddIndexType;
+            var indexType = Configurations.ProductRemoveIndexType;
 
             await elasticClient.IndexAsync(productModel,
                 i => i.Index(defaultIndex).Type(indexType).Refresh());
