+++ OnlinerTask/OnlinerTask.Data/OnlinerTask.DAL.csproj
@@ -30,6 +30,10 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="Elasticsearch.Net, Version=2.0.0.0, Culture=neutral, PublicKeyToken=96c599bbe3e70f5d, processorArchitecture=MSIL">
+      <HintPath>..\packages\Elasticsearch.Net.2.4.7\lib\net45\Elasticsearch.Net.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
     <Reference Include="EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089, processorArchitecture=MSIL">
       <HintPath>..\packages\EntityFramework.6.1.3\lib\net45\EntityFramework.dll</HintPath>
       <Private>True</Private>
@@ -55,6 +59,10 @@
     </Reference>
     <Reference Include="Microsoft.Web.Infrastructure, Version=1.0.0.0, Culture=neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=MSIL">
       <HintPath>..\packages\Microsoft.Web.Infrastructure.1.0.0.0\lib\net40\Microsoft.Web.Infrastructure.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
+    <Reference Include="Nest, Version=2.0.0.0, Culture=neutral, PublicKeyToken=96c599bbe3e70f5d, processorArchitecture=MSIL">
+      <HintPath>..\packages\NEST.2.4.7\lib\net45\Nest.dll</HintPath>
       <Private>True</Private>
     </Reference>
     <Reference Include="NetMQ, Version=3.3.0.11, Culture=neutral, PublicKeyToken=a6decef4ddc58b3a, processorArchitecture=MSIL">
@@ -163,6 +171,13 @@
     <Compile Include="DataBaseModels\Review.cs" />
     <Compile Include="DataBaseModels\UpdatedProducts.cs" />
     <Compile Include="DataStartUp.cs" />
+    <Compile Include="ElasticSearch\ClientsFabric\IClientsFactory.cs" />
+    <Compile Include="ElasticSearch\ClientsFabric\Implementations\ElasticClientsFactory.cs" />
+    <Compile Include="ElasticSearch\ConnectionFabric\IConnectionFactory.cs" />
+    <Compile Include="ElasticSearch\ConnectionFabric\Implementations\ElasticConnectionFactory.cs" />
+    <Compile Include="ElasticSearch\ProductLogger\Implementations\ProductAddLogger.cs" />
+    <Compile Include="ElasticSearch\ProductLogger\Implementations\ProductRemoveLogger.cs" />
+    <Compile Include="ElasticSearch\ProductLogger\IProductLogger.cs" />
     <Compile Include="EntityMappers\IDependentMapper.cs" />
     <Compile Include="EntityMappers\IMapper.cs" />
     <Compile Include="EntityMappers\Implementations\ImageMapper.cs" />
