+++ OnlinerTask/OnlinerTask.Data/packages.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Elasticsearch.Net" version="2.4.7" targetFramework="net452" />
   <package id="EntityFramework" version="6.1.3" targetFramework="net452" />
   <package id="jQuery" version="1.6.4" targetFramework="net452" />
   <package id="Microsoft.AspNet.Identity.Core" version="2.2.1" targetFramework="net452" />
@@ -15,6 +16,7 @@
   <package id="Microsoft.Owin.Host.SystemWeb" version="2.1.0" targetFramework="net452" />
   <package id="Microsoft.Owin.Security" version="2.1.0" targetFramework="net452" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net452" />
+  <package id="NEST" version="2.4.7" targetFramework="net452" />
   <package id="NetMQ" version="3.3.0.11" targetFramework="net452" />
   <package id="NetMQ.WebSockets" version="0.5.0.0-beta" targetFramework="net452" />
   <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net452" />
