+++ OnlinerTask/OnlinerTask.WEB/App_Start/NinjectWebCommon.cs
@@ -1,12 +1,6 @@
 using Ninject.Syntax;
 using OnlinerTask.BLL.Services.ConfigChange;
 using OnlinerTask.BLL.Services.ConfigChange.Implementations;
-using OnlinerTask.BLL.Services.ElasticSearch.ClientsFabric;
-using OnlinerTask.BLL.Services.ElasticSearch.ClientsFabric.Implementations;
-using OnlinerTask.BLL.Services.ElasticSearch.ConnectionFabric;
-using OnlinerTask.BLL.Services.ElasticSearch.ConnectionFabric.Implementations;
-using OnlinerTask.BLL.Services.ElasticSearch.ProductLogger;
-using OnlinerTask.BLL.Services.ElasticSearch.ProductLogger.Implementations;
 using OnlinerTask.BLL.Services.Job.EmailJob;
 using OnlinerTask.BLL.Services.Job.EmailJob.Implementations;
 using OnlinerTask.BLL.Services.Job.ProductJob;
@@ -29,6 +23,12 @@ using OnlinerTask.BLL.Services.TimeChange;
 using OnlinerTask.BLL.Services.TimeChange.Implementations;
 using OnlinerTask.Data.DataBaseContexts;
 using OnlinerTask.Data.DataBaseInterfaces;
+using OnlinerTask.Data.ElasticSearch.ClientsFabric;
+using OnlinerTask.Data.ElasticSearch.ClientsFabric.Implementations;
+using OnlinerTask.Data.ElasticSearch.ConnectionFabric;
+using OnlinerTask.Data.ElasticSearch.ConnectionFabric.Implementations;
+using OnlinerTask.Data.ElasticSearch.ProductLogger;
+using OnlinerTask.Data.ElasticSearch.ProductLogger.Implementations;
 using OnlinerTask.Data.EntityMappers.Implementations;
 using OnlinerTask.Data.IdentityModels;
 using OnlinerTask.Data.MqConstituents;
